# Feature #612 - implement LogFilter glob-like wildcard support