--- a/core/jvm/src/test/scala/zio/logging/LogFilterSpec.scala
+++ b/core/jvm/src/test/scala/zio/logging/LogFilterSpec.scala
@@ -88,11 +88,11 @@ object LogFilterSpec extends ZIOSpecDefault {
         "a"         -> LogLevel.Info,
         "a.b.c"     -> LogLevel.Warning,
         "e.f"       -> LogLevel.Error,
-        "k.*.m"     -> LogLevel.Trace,
-        "k2.a*c.m2" -> LogLevel.Trace,
-        "k3.a*.m3"  -> LogLevel.Trace,
-        "k4.*c.m4"  -> LogLevel.Trace,
-        "q.**.t"    -> LogLevel.Trace
+        "k.*.m"     -> LogLevel.Info,
+        "k2.a*c.m2" -> LogLevel.Info,
+        "k3.a*.m3"  -> LogLevel.Info,
+        "k4.*c.m4"  -> LogLevel.Info,
+        "q.**.t"    -> LogLevel.Info
       )
 
       testFilter(filter, "x.Exec.exec", LogLevel.Debug, Assertion.isTrue) &&
@@ -105,17 +105,20 @@ object LogFilterSpec extends ZIOSpecDefault {
       testFilter(filter, "e.Exec.exec", LogLevel.Debug, Assertion.isTrue) &&
       testFilter(filter, "e.f.Exec.exec", LogLevel.Debug, Assertion.isFalse) &&
       testFilter(filter, "k.l.Exec.exec", LogLevel.Debug, Assertion.isTrue) &&
-      testFilter(filter, "k.l.m.Exec.exec", LogLevel.Trace, Assertion.isTrue) &&
-      testFilter(filter, "k2.alc.m2.Exec.exec", LogLevel.Trace, Assertion.isTrue) &&
-      testFilter(filter, "k3.alc.m3.Exec.exec", LogLevel.Trace, Assertion.isTrue) &&
-      testFilter(filter, "k3.lc.m3.Exec.exec", LogLevel.Trace, Assertion.isFalse) &&
-      testFilter(filter, "k4.alc.m4.Exec.exec", LogLevel.Trace, Assertion.isTrue) &&
-      testFilter(filter, "k4.al.m4.Exec.exec", LogLevel.Trace, Assertion.isFalse) &&
-      testFilter(filter, "k.l.l.m.Exec.exec", LogLevel.Trace, Assertion.isFalse) &&
+      testFilter(filter, "k.l.m.Exec.exec", LogLevel.Debug, Assertion.isFalse) &&
+      testFilter(filter, "k.l.m.Exec.exec", LogLevel.Info, Assertion.isTrue) &&
       testFilter(filter, "k.l.l.m.Exec.exec", LogLevel.Debug, Assertion.isTrue) &&
-      testFilter(filter, "q.r.s.t.Exec.exec", LogLevel.Trace, Assertion.isTrue) &&
-      testFilter(filter, "q.r.t.Exec.exec", LogLevel.Trace, Assertion.isTrue) &&
-      testFilter(filter, "q.r.s.u.Exec.exec", LogLevel.Trace, Assertion.isFalse) &&
+      testFilter(filter, "k2.alc.m2.Exec.exec", LogLevel.Info, Assertion.isTrue) &&
+      testFilter(filter, "k3.alc.m3.Exec.exec", LogLevel.Info, Assertion.isTrue) &&
+      testFilter(filter, "k3.alc.m3.Exec.exec", LogLevel.Warning, Assertion.isTrue) &&
+      testFilter(filter, "k3.lc.m3.Exec.exec", LogLevel.Debug, Assertion.isTrue) &&
+      testFilter(filter, "k4.alc.m4.Exec.exec", LogLevel.Debug, Assertion.isFalse) &&
+      testFilter(filter, "k4.alc.m4.Exec.exec", LogLevel.Info, Assertion.isTrue) &&
+      testFilter(filter, "k4.al.m4.Exec.exec", LogLevel.Debug, Assertion.isTrue) &&
+      testFilter(filter, "q.r.t.Exec.exec", LogLevel.Debug, Assertion.isFalse) &&
+      testFilter(filter, "q.r.t.Exec.exec", LogLevel.Info, Assertion.isTrue) &&
+      testFilter(filter, "q.r.s.t.Exec.exec", LogLevel.Debug, Assertion.isFalse) &&
+      testFilter(filter, "q.r.s.t.Exec.exec", LogLevel.Info, Assertion.isTrue) &&
       testFilter(filter, "q.r.s.u.Exec.exec", LogLevel.Debug, Assertion.isTrue)
     },
     test("log filtering by log level and name with annotation") {