--- a/core/shared/src/main/scala/zio/logging/LogFilter.scala
+++ b/core/shared/src/main/scala/zio/logging/LogFilter.scala
@@ -284,10 +284,54 @@ object LogFilter {
     val mappingsSorted = mappings.map(splitNameByDotAndLevel.tupled).sorted(nameLevelOrdering)
     val nameGroup      = group.map(splitNameByDot)
 
+    @tailrec
+    def compareRoutes(l: List[String], m: List[String]): Boolean = (l, m) match {
+      case (Nil, Nil)                       => true
+      case (Nil, _)                         => false
+      case (l, Nil)                         => true
+      case (l @ (lh :: ls), m @ (mh :: ms)) =>
+        l.startsWith(m) ||
+          (m.span(g => !g.contains("*")) match {
+            case (_, Nil)                => false
+            case (m0, "**" :: Nil)       => true
+            case (Nil, "*" :: m)         =>
+              // we drop lh which is matched by '*'
+              compareRoutes(ls, m)
+            case (m0, "*" :: m)          =>
+              // match the m0 part and remove it from the list before the compare the remainder
+              l.startsWith(m0) && (l.drop(m0.size) match {
+                case Nil      => false
+                case lh :: ls => compareRoutes(ls, m)
+              })
+            case (Nil, "**" :: mh :: ms) =>
+              // remove any paths not matching the mh part
+              l.dropWhile(_ != mh) match {
+                case Nil      => false // we didn't find a match so this route is invalid
+                case lh :: ls => compareRoutes(ls, ms)
+              }
+            case (m0, "**" :: mh :: ms)  =>
+              // match the m0 part and remove it from the list before the compare the remainder
+              l.startsWith(m0) && (
+                ls.dropWhile(_ != mh) match {
+                  case Nil      => false
+                  case lh :: ls => compareRoutes(ls, ms)
+                }
+              )
+            case (m0, s"$s*$e" :: m)     =>
+              // match the m0 part and remove it from the list before the compare the remainder
+              l.startsWith(m0) && (l.drop(m0.size) match {
+                case Nil      => false
+                case lh :: ls =>
+                  // check if lh starts with s, strip it (we don't want overlap) and ends with e, then the compare the remainder
+                  lh.startsWith(s) && lh.stripPrefix(s).endsWith(e) && compareRoutes(ls, m)
+              })
+          })
+    }
+
     logLevelByGroup[M, List[String]](
       rootLevel,
       nameGroup,
-      (l, m) => l.startsWith(m),
+      (l, m) => compareRoutes(l, m),
       mappingsSorted: _*
     )
   }