#612

I implemented `**` (any number of paths) and `*` (any (partial) path) within the existing LogFilter path structure `List[String]`. 

I think this could be optimized if the LogFilters were used in a nested map-like structure. 

Another way would be the use regex.