@ThijsBroersen 2023-01-08
Did some benchmarks to see if there is much performance loss.

Feature #612
```
[info] Benchmark                                           Mode  Cnt     Score       Error  Units
[info] FilterBenchmarks.cachedFilterByLogLevelAndNameLog  thrpt    3  5668.719 ± 39326.263  ops/s
[info] FilterBenchmarks.filterByLogLevelAndNameLog        thrpt    3  4649.646 ± 12339.049  ops/s
[info] FilterBenchmarks.handWrittenFilterLog              thrpt    3  3109.203 ±  2338.471  ops/s
[info] FilterBenchmarks.noFilteringLog                    thrpt    3  2270.438 ±  8116.408  ops/s```

Master
```[info] Benchmark                                           Mode  Cnt     Score       Error  Units
[info] FilterBenchmarks.cachedFilterByLogLevelAndNameLog  thrpt    3  5390.427 ± 35953.429  ops/s
[info] FilterBenchmarks.filterByLogLevelAndNameLog        thrpt    3  4323.591 ± 22497.521  ops/s
[info] FilterBenchmarks.handWrittenFilterLog              thrpt    3  2725.011 ±   348.357  ops/s
[info] FilterBenchmarks.noFilteringLog                    thrpt    3  1957.494 ±  1254.506  ops/s```