@ThijsBroersen 2023-01-08
Hmm, I see this library still needs to support older Scala versions for which the CI tests now fail 😞