@ThijsBroersen 2023-03-25
PR
```
[info] Benchmark                                           Mode  Cnt      Score      Error  Units
[info] FilterBenchmarks.cachedFilterByLogLevelAndNameLog  thrpt    3  11606.209 ± 4034.046  ops/s
[info] FilterBenchmarks.filterByLogLevelAndNameLog        thrpt    3  11465.383 ± 7223.445  ops/s
[info] FilterBenchmarks.handWrittenFilterLog              thrpt    3   9420.996 ± 1500.501  ops/s
[info] FilterBenchmarks.noFilteringLog                    thrpt    3   8159.752 ± 1312.800  ops/s
```
Master
```
[info] Benchmark                                           Mode  Cnt      Score      Error  Units
[info] FilterBenchmarks.cachedFilterByLogLevelAndNameLog  thrpt    3  11525.222 ± 1802.478  ops/s
[info] FilterBenchmarks.filterByLogLevelAndNameLog        thrpt    3  11873.502 ± 2359.884  ops/s
[info] FilterBenchmarks.handWrittenFilterLog              thrpt    3   9522.591 ±  927.634  ops/s
[info] FilterBenchmarks.noFilteringLog                    thrpt    3   7957.346 ± 1085.727  ops/s
```